refactor: rename _output_mixin.py → _output.py
Now pure output (render/compile dispatch + validation). Schema
introspection moved to _grammar.py in previous commit.
